chore(sdk-coin-avaxc): move bitgo source code
- moves code from bitgo to sdk-coin-avaxc

Ticket: BG-48144
