--- modules/bitgo/src/v2/coins/tavaxc.ts
@@ -1,17 +1,0 @@
-/**
- * @prettier
- */
-import { BaseCoin, BitGoBase } from '@bitgo/sdk-core';
-import { BaseCoin as StaticsBaseCoin } from '@bitgo/statics';
-
-import { AvaxC } from './avaxc';
-
-export class TavaxC extends AvaxC {
-  protected constructor(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>) {
-    super(bitgo, staticsCoin);
-  }
-
-  static createInstance(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>): BaseCoin {
-    return new TavaxC(bitgo, staticsCoin);
-  }
-}
