+++ modules/sdk-coin-avaxc/package.json
@@ -41,7 +41,14 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/sdk-coin-eth": "^1.0.0",
+    "@bitgo/statics": "^6.17.0",
+    "bignumber.js": "^8.0.1",
+    "bip32": "^2.0.6",
+    "keccak": "^3.0.2",
+    "lodash": "^4.17.14",
+    "secp256k1": "^4.0.2"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.0.1",
