+++ modules/sdk-coin-avaxc/src/avaxcToken.ts
