+++ modules/sdk-coin-avaxc/src/index.ts
@@ -1,2 +1,3 @@
 export * from './avaxc';
+export * from './avaxcToken';
 export * from './tavaxc';
